Add get_incidents to CoreIR (#31333)
* initial commit

* fixes

* some more fixes

* some more fixes

* adding release notes

* fixes

* added command to CoreIR readme

* more fixes

* more fixes

* more fixes

* Bump pack from version Core to 3.0.4.

* Bump pack from version CortexXDR to 6.0.9.

* unifying implementation

* adding release notes

* adding release notes

* Bump pack from version DemistoRESTAPI to 1.3.42.

* Bump pack from version CortexXDR to 6.0.10.

* Bump pack from version ctf01 to 1.0.4.

* Bump pack from version CortexXDR to 6.0.11.

* release notes

* docker updates

* Bump pack from version DemistoRESTAPI to 1.3.44.

* release notes

* update release notes

* Update configure_and_test_integration_instances.py

testing marketplace sync time

* Bump pack from version CortexXDR to 6.1.2.

* triggering upload and install

* reverting changes to CoreRESTAPI to see if it fixes the build

* Bump pack from version CortexXDR to 6.1.3.

* Bump pack from version ctf01 to 1.0.5.

* release notes

* Bump pack from version Core to 3.0.7.

* reverting change

* release notes change

* Bump pack from version Core to 3.0.8.

* revert changes in ctf

* revert changes in ctf

* updating docker images

* deleting from ctf again

* deleting from ctf again

* updating release notes

* reverting changes

* Bump pack from version Core to 3.0.11.

* Bump pack from version Core to 3.0.12.

* updating docker image

* Bump pack from version Core to 3.0.13.

* Bump pack from version CortexXDR to 6.1.6.

---------
