+++ Packs/ApiModules/ReleaseNotes/2_2_21.md
@@ -1,0 +1,4 @@
+
+#### Scripts
+##### CoreIRApiModule
+- Improved implementation of *xdr-get-incidents* by unifying similar code blocks.
