+++ Packs/Core/Integrations/CortexCoreIR/CortexCoreIR.py
@@ -464,6 +464,9 @@ def main():  # pragma: no cover
         elif command == 'core-list-roles':
             return_results(list_roles_command(client, args))
 
+        elif command == 'core-get-incidents':
+            return_outputs(*get_incidents_command(client, args))
+
         elif command in ('core-set-user-role', 'core-remove-user-role'):
             return_results(change_user_role_command(client, args))
