+++ Packs/Core/ReleaseNotes/3_0_13.md
@@ -1,0 +1,7 @@
+
+#### Integrations
+
+##### Investigation & Response
+
+- Added the *core-get-incidents* command.
+- Updated the Docker image to: *demisto/python3:3.10.13.84405*.
