+++ Packs/CortexXDR/Integrations/CortexXDRIR/CortexXDRIR.yml
@@ -3472,7 +3472,7 @@ script:
       isArray: true
     name: xdr-remove-user-role
     description: Remove one or more users from a role.
-  dockerimage: demisto/python3:3.10.13.83255
+  dockerimage: demisto/python3:3.10.13.84405
   isfetch: true
   isfetch:xpanse: false
   script: ''
