+++ Packs/CortexXDR/ReleaseNotes/6_1_6.md
@@ -1,0 +1,7 @@
+
+#### Integrations
+
+##### Palo Alto Networks Cortex XDR - Investigation and Response
+
+- Improved implementation of *xdr-get-incidents* by unifying similar code blocks.
+- Updated the Docker image to: *demisto/python3:3.10.13.84405*.
