+++ Packs/ctf01/ReleaseNotes/1_0_6.md
@@ -1,0 +1,7 @@
+
+#### Integrations
+
+##### Cortex XDR - IR CTF
+
+- Improved implementation by removing redundant code blocks.
+- Updated the Docker image to: *demisto/python3:3.10.13.84405*.
